Move loading function in resources
