+++ snips-nlu-lib/src/intent_parser/deterministic_intent_parser.rs
@@ -339,7 +339,7 @@ mod tests {
     use super::*;
 
     use models::DeterministicParserModel;
-    use nlu_engine::load_engine_shared_resources;
+    use resources::loading::load_engine_shared_resources;
     use slot_utils::InternalSlot;
     use snips_nlu_ontology::*;
     use testutils::*;
